Chest Loot and Exit Key
I added in loot that spawns from the chest when opened and player to be able to 'pick' it up. Currently it isn't stored anywhere waiting for the frontend side to catch up. The enemy now drops a Key which is used to unlock the exit tile which will be used to progress further.