Linking Up Scenes
Linked up the main menu, game and lose scene together. When the players hp reaches 0 it plays the death animation and the player is taken to the lose scene which doesn't have any UI working for it other than 1 of the button which takes the player back to the main menu where they can select start again to player again in a new run. Also tried fixing the shield UI as currently its a bit broken meaning when loading into the next level it doesn't load correctly same as when the use restarts a game.